Consolidate server form use cases
